+++ lib/versioncake/strategies/http_header_strategy.rb
@@ -2,8 +2,8 @@ module VersionCake
   class HttpHeaderStrategy < ExtractionStrategy
 
     def execute(request)
-      if request.headers.key? "HTTP_X_#{version_key.upcase}"
-        request.headers["HTTP_X_#{version_key.upcase}"]
+      if request.headers.key? "HTTP_#{version_key.upcase}"
+        request.headers["HTTP_#{version_key.upcase}"]
       end
     end