+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Bug Fixes:
 
 * When an invalid version is received an unsupported exception is raised (#34)
+* Remove deprecated X- from header
 
 Enhancements: